refactor(Direction, Point): domain 패키지로 이동
