+++ src/main/java/laddergame/domain/Line.java
@@ -1,7 +1,5 @@
 package laddergame.domain;
 
-import laddergame.Point;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
