+++ src/test/java/laddergame/domain/PointTest.java
@@ -1,4 +1,4 @@
-package laddergame;
+package laddergame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
